+++ changelog.md
@@ -4,6 +4,7 @@ TGUI 1.8  (TBD)
 - New widget: ContextMenu
 - Added handleWindowEvents function to SFML backend (only for SFML 3)
 - Added onRightClick signal to ListBox
+- Added onTabRightClick signal to Tabs
 - Added setSmooth and isSmooth to CanvasSFML
 - Changed the way the view is handled in CanvasSFML
 - Pressing escape now closes the menu from ContextMenu or MenuBar